pyproject.toml
[project]
name = "homeproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "asyncpg>=0.30.0",
    "bcrypt==4.3.0",
    "black>=25.9.0",
    "codespell>=2.4.1",
    "fastapi>=0.119.1",
    "fastapi-pagination>=0.14.3",
    "freezegun>=1.5.5",
    "httpx>=0.28.1",
    "isort>=7.0.0",
    "mypy>=1.18.2",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic[email]>=2.12.3",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "python-jose>=3.5.0",
    "refurb>=2.2.0",
    "ruff>=0.14.2",
    "sqlalchemy>=2.0.44",
    "strawberry-graphql[fastapi]>=0.284.1",
    "types-passlib>=1.7.7.20250602",
    "types-python-jose>=3.5.0.20250531",
    "uvicorn>=0.38.0",
]

[tool.mypy]
plugins = ["strawberry.ext.mypy_plugin"]
